fix: resolve `ctx.{currentStep}` being `undefined` in `createComponent` with custom `useFormInstance` and no `ctxData` #102
